refactor: 승패 여부 계산 로직 분리
- Judgement라는 클래스로 로직을 이동시켰다.
- GameResult를 JudgementResult로 변경했다.
