+++ src/test/java/blackjack/domain/DealerTest.java
@@ -79,153 +79,4 @@ class DealerTest {
 
         assertThat(result).containsExactly(card2);
     }
-
-    @DisplayName("플레이어와 딜러가 둘 다 버스트되면, 플레이어가 진다.")
-    @Test
-    void allBust() {
-        Dealer dealer = new Dealer();
-        dealer.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-        dealer.hit(new Card(CardRank.FOUR, CardShape.DIAMOND));
-        dealer.hit(new Card(CardRank.QUEEN, CardShape.DIAMOND));
-
-        Player player = new Player("pobi", new Money(1000));
-        player.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-        player.hit(new Card(CardRank.JACK, CardShape.DIAMOND));
-        player.hit(new Card(CardRank.QUEEN, CardShape.DIAMOND));
-
-        GameResult gameResult = dealer.judge(player);
-
-        assertThat(gameResult).isEqualTo(GameResult.LOSE);
-    }
-
-    @DisplayName("플레이어와 딜러가 둘 다 블랙잭이면, 무승부이다.")
-    @Test
-    void allBlackJack() {
-        Dealer dealer = new Dealer();
-        dealer.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-        dealer.hit(new Card(CardRank.ACE, CardShape.DIAMOND));
-
-        Player player = new Player("pobi", new Money(1000));
-        player.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-        player.hit(new Card(CardRank.ACE, CardShape.DIAMOND));
-
-        GameResult gameResult = dealer.judge(player);
-
-        assertThat(gameResult).isEqualTo(GameResult.TIE);
-    }
-
-    @DisplayName("딜러만 블랙잭인 경우, 플레이어가 진다.")
-    @Test
-    void onlyDealerBlackJack() {
-        Dealer dealer = new Dealer();
-        dealer.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-        dealer.hit(new Card(CardRank.ACE, CardShape.DIAMOND));
-
-        Player player = new Player("pobi", new Money(1000));
-        player.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-        player.hit(new Card(CardRank.QUEEN, CardShape.DIAMOND));
-        player.hit(new Card(CardRank.ACE, CardShape.DIAMOND));
-
-        GameResult gameResult = dealer.judge(player);
-
-        assertThat(gameResult).isEqualTo(GameResult.LOSE);
-    }
-
-    @DisplayName("플레이어만 블랙잭인 경우, 딜러가 진다.")
-    @Test
-    void onlyPlayerBlackJack() {
-        Dealer dealer = new Dealer();
-        dealer.hit(new Card(CardRank.SIX, CardShape.DIAMOND));
-        dealer.hit(new Card(CardRank.FIVE, CardShape.DIAMOND));
-        dealer.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-
-        Player player = new Player("pobi", new Money(1000));
-        player.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-        player.hit(new Card(CardRank.ACE, CardShape.DIAMOND));
-
-        GameResult gameResult = dealer.judge(player);
-
-        assertThat(gameResult).isEqualTo(GameResult.BLACKJACK_WIN);
-    }
-
-    @DisplayName("딜러만 버스트되면, 플레이어가 이긴다.")
-    @Test
-    void onlyDealerBust() {
-        Dealer dealer = new Dealer();
-        dealer.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-        dealer.hit(new Card(CardRank.FOUR, CardShape.DIAMOND));
-        dealer.hit(new Card(CardRank.QUEEN, CardShape.DIAMOND));
-
-        Player player = new Player("pobi", new Money(1000));
-        player.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-
-        GameResult gameResult = dealer.judge(player);
-
-        assertThat(gameResult).isEqualTo(GameResult.WIN);
-    }
-
-    @DisplayName("플레이어만 버스트되면, 딜러가 이긴다.")
-    @Test
-    void onlyPlayerBust() {
-        Dealer dealer = new Dealer();
-        dealer.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-        dealer.hit(new Card(CardRank.QUEEN, CardShape.DIAMOND));
-
-        Player player = new Player("pobi", new Money(1000));
-        player.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-        player.hit(new Card(CardRank.FOUR, CardShape.DIAMOND));
-        player.hit(new Card(CardRank.QUEEN, CardShape.DIAMOND));
-
-        GameResult gameResult = dealer.judge(player);
-
-        assertThat(gameResult).isEqualTo(GameResult.LOSE);
-    }
-
-    @DisplayName("플레이어와 딜러가 둘 다 버스트되지 않고, 딜러 점수가 더 낮으면 플레이어가 이긴다.")
-    @Test
-    void whenDealerLose() {
-        Dealer dealer = new Dealer();
-        dealer.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-        dealer.hit(new Card(CardRank.SEVEN, CardShape.DIAMOND));
-
-        Player player = new Player("pobi", new Money(1000));
-        player.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-        player.hit(new Card(CardRank.ACE, CardShape.DIAMOND));
-
-        GameResult gameResult = dealer.judge(player);
-
-        assertThat(gameResult).isEqualTo(GameResult.WIN);
-    }
-
-    @DisplayName("플레이어와 딜러가 둘 다 버스트되지 않고, 딜러 점수가 더 높으면 플레이어가 진다.")
-    @Test
-    void whenDealerWin() {
-        Dealer dealer = new Dealer();
-        dealer.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-        dealer.hit(new Card(CardRank.ACE, CardShape.DIAMOND));
-
-        Player player = new Player("pobi", new Money(1000));
-        player.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-        player.hit(new Card(CardRank.SEVEN, CardShape.DIAMOND));
-
-        GameResult gameResult = dealer.judge(player);
-
-        assertThat(gameResult).isEqualTo(GameResult.LOSE);
-    }
-
-    @DisplayName("플레이어와 딜러가 둘 다 버스트되지 않고, 같은 점수라면 무승부다.")
-    @Test
-    void whenTie() {
-        Dealer dealer = new Dealer();
-        dealer.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-        dealer.hit(new Card(CardRank.QUEEN, CardShape.DIAMOND));
-
-        Player player = new Player("pobi", new Money(1000));
-        player.hit(new Card(CardRank.KING, CardShape.DIAMOND));
-        player.hit(new Card(CardRank.QUEEN, CardShape.DIAMOND));
-
-        GameResult gameResult = dealer.judge(player);
-
-        assertThat(gameResult).isEqualTo(GameResult.TIE);
-    }
 }
